Move platform util package
