+++ pkg/util/platform/platform.go
@@ -1,4 +1,4 @@
-package util
+package platform
 
 import (
 	"context"
