+++ pkg/util/platform/platform_test.go
@@ -1,4 +1,4 @@
-package util
+package platform
 
 import (
 	"context"
